requirements.txt
fastapi
requests
Pillow
twilio
schedule
easyocr
numpy
openai-whisper
langchain
langchain-community
langchain-huggingface
faiss-cpu
gTTS
python-dotenv
streamlit
pandas
plotly
pydantic
uvicorn
